Optimize genprimes

+++ src/lab1/lab1.c
@@ -2,22 +2,22 @@
 #include <stdio.h>
 #include "lab1.h"
 
-void lab1_count_factors(int min, int max, int count, int results[], ...)
+void lab1_count_factors(long min, long max, long count, long results[], ...)
 {
-    for (int j = 0; j < count; j++)
+    for (long j = 0; j < count; j++)
     {
         results[j] = 0;
     }
 
     va_list params;
 
-    for (int i = min; i <= max; i++)
+    for (long i = min; i <= max; i++)
     {
         va_start(params, results);
 
-        for (int j = 0; j < count; j++)
+        for (long j = 0; j < count; j++)
         {
-            if (i % va_arg(params, int) == 0)
+            if (i % va_arg(params, long) == 0)
             {
                 results[j]++;
             }
@@ -27,15 +27,15 @@ void lab1_count_factors(int min, int max, int count, int results[], ...)
     }
 }
 
-void lab1_print_factors(int count, int results[], ...)
+void lab1_print_factors(long count, long results[], ...)
 {
     va_list params;
 
     va_start(params, results);
 
-    for (int j = 0; j < count; j++)
+    for (long j = 0; j < count; j++)
     {
-        printf("%d: %d\n", va_arg(params, int), results[j]);
+        printf("%ld: %ld\n", va_arg(params, long), results[j]);
     }
 
     va_end(params);
